requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.4.0
audioop-lts==0.2.2
discord.py==2.6.4
dotenv==0.9.9
frozenlist==1.8.0
idna==3.11
multidict==6.7.0
orjson==3.11.5
pomice==2.10.0
propcache==0.4.1
python-dotenv==1.2.1
typing_extensions==4.15.0
websockets==15.0.1
yarl==1.22.0
